feat: create specific file to filterList
